获取用户已看过电影的电影信息及评价信息接口实现。🎼

+++ src/main/java/cn/zjw/mrs/entity/Movie.java
@@ -25,7 +25,7 @@ public class Movie implements Serializable {
     /**
      * 豆瓣ID
      */
-    private Integer did;
+    private Long did;
 
     /**
      * 电影名
